JOTM-26/10/2018-Ventas
* Servicio de ventas

+++ 1.Layers/1.3Application/1.3.2MainModule/Application.MainModule/DTOs/Mobile/VentaDTO.cs
@@ -9,7 +9,7 @@ namespace Application.MainModule.DTOs.Mobile
     public class VentaDTO
     {
         public string FolioVenta { get; set; }
-        public short IdCliente { get; set; }
+        public int IdCliente { get; set; }
         public decimal Subtotal { get; set; }
         public decimal Iva { get; set; }
         public decimal Total { get; set; }
